# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TRAMPO — Plataforma de matchmaking de carreiras com IA
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: Listar vagas
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: work_mode
          in: query
          schema:
            type: string
            enum: [remote, hybrid, onsite, any]
        - name: contract_type
          in: query
          schema:
            type: string
            enum: [clt, pj, freelancer, internship]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Lista de vagas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Publicar vaga (recrutador)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Vaga criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Ver vaga específica
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vaga
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    put:
      operationId: updateJob
      tags: [jobs]
      summary: Editar vaga
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "200":
          description: Vaga atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Deletar vaga
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vaga deletada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /payments/status:
    get:
      operationId: getPaymentStatus
      tags: [payments]
      summary: Ver se sou premium
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Status do plano
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStatus"
  /payments/create-checkout:
    post:
      operationId: createCheckout
      tags: [payments]
      summary: Criar sessão de pagamento
      security:
        - bearerAuth: []
      responses:
        "200":
          description: URL de checkout
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkout_url:
                    type: string
                required: [checkout_url]
  /support/tickets:
    get:
      operationId: listSupportTickets
      tags: [support]
      summary: Listar meus tickets
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createSupportTicket
      tags: [support]
      summary: Criar ticket de suporte
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportTicketRequest"
      responses:
        "201":
          description: Ticket criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
...
  /recruiters/jobs:
    get:
      operationId: getRecruiterJobs
      tags: [recruiters]
      summary: Vagas do recrutador
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Vagas do recrutador
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Job:
      type: object
      properties:
        id:
          type: integer
        recruiter_id:
          type: integer
          nullable: true
        title:
          type: string
        company:
          type: string
        description:
          type: string
        location:
          type: string
          nullable: true
        cidade:
          type: string
          nullable: true
        estado:
          type: string
          nullable: true
        contract_type:
          type: string
          enum: [clt, pj, freelancer, internship]
          nullable: true
        salary_min:
          type: number
          nullable: true
        salary_max:
          type: number
          nullable: true
        work_mode:
          type: string
          enum: [remote, hybrid, onsite]
          nullable: true
        required_skills:
          type: string
          nullable: true
        required_experience:
          type: integer
          nullable: true
        contact_email:
          type: string
          nullable: true
        benefits:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, closed]
        source:
          type: string
          nullable: true
        is_featured:
          type: boolean
        posted_at:
          type: string
          format: date-time
      required: [id, title, company, description, status, is_featured, posted_at]
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        total_pages:
          type: integer
      required: [jobs, total, page, total_pages]
    CreateJobRequest:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        description:
          type: string
        location:
          type: string
        cidade:
          type: string
        estado:
          type: string
        contract_type:
          type: string
          enum: [clt, pj, freelancer, internship]
        salary_min:
          type: number
        salary_max:
          type: number
        work_mode:
          type: string
          enum: [remote, hybrid, onsite]
        required_skills:
          type: string
        required_experience:
          type: integer
        contact_email:
          type: string
        benefits:
          type: string
      required: [title, company, description]
...
    PaymentStatus:
      type: object
      properties:
        is_premium:
          type: boolean
        premium_expires_at:
          type: string
          format: date-time
          nullable: true
        daily_sends_used:
          type: integer
        daily_limit:
          type: integer
      required: [is_premium, daily_sends_used, daily_limit]
    SupportTicket:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [open, resolved]
        priority:
          type: string
          enum: [normal, high]
        created_at:
          type: string
          format: date-time
      required: [id, user_id, subject, message, status, priority, created_at]
    CreateSupportTicketRequest:
      type: object
      properties:
        subject:
          type: string
        message:
          type: string
      required: [subject, message]
...