# lib/api-zod/src/generated/types/createJobRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TRAMPO — Plataforma de matchmaking de carreiras com IA
 * OpenAPI spec version: 0.1.0
 */
import type { CreateJobRequestContractType } from "./createJobRequestContractType";
import type { CreateJobRequestWorkMode } from "./createJobRequestWorkMode";

export interface CreateJobRequest {
  title: string;
  company: string;
  description: string;
  location?: string;
  cidade?: string;
  estado?: string;
  contract_type?: CreateJobRequestContractType;
  salary_min?: number;
  salary_max?: number;
  work_mode?: CreateJobRequestWorkMode;
  required_skills?: string;
  required_experience?: number;
  contact_email?: string;
  benefits?: string;
}