# lib/api-zod/src/generated/types/createSupportTicketRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TRAMPO — Plataforma de matchmaking de carreiras com IA
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSupportTicketRequest {
  subject: string;
  message: string;
}